Allow dynamic external function values
Whenever the function was being bound (we can't guarantee which NPC binds the external function), it was setting the vectors equal to THAT object's values. However, if a different object called the external, the values would then be wrong. Now using the object the player is interacting with to get the correct vector, dynamically.