Prevent dialogue window being hidden if player starts a new conversation
Added a boolean variable. Every time the player initiates a conversation, the boolean is set to true.

The reason for this is that a coroutine begins a process of closing the dialogue window once the conversation ends; however, if the player starts a new conversation before this timer runs out, it will close the dialogue window when the player is in a new conversation.

The fix is setting a boolean at the start and end of the conversation. The display window is only closed if this value is false. (I.e., the player hasn't attempted a new conversation since the last one ended.)