Make Keybind system more dynamic
- Dict 1: Key/Abil
- Dict 2: Abil/Action

These two dictionaries are set at runtime or with loads. They can reference each other to get information. The "Abil" ties them both together. For example, if you have a keybind, you can get the ability, and then use that ability to perform an action.

Static setter method in GameMetaInfo to set values for the dictionary.