Fix Enemy Stats on Spawn
Enemies subscribed to game load event for setting stats - but delay on spawner means this method is never called.

The method is now called in start, as well as subscribed. To prevent the stats being compounded, the stats are reset to their default value each time before applying other modifiers.